% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{americas_clip}
\alias{americas_clip}
\title{Clipping polygon for the Americas}
\format{
\subsection{\code{americas_clip}}{

A \code{sfc} object as used by the \strong{sf} package. This is used to define
North and South America when constraining BirdFlow models to just the
Americas.  It is made by mergin together all the Natural Earth countries
that are in the Amercas, dropping Hawaii, and then buffering by 300 km.
}
}
\usage{
americas_clip
}
\description{
Clipping polygon for the Americas
}
\keyword{datasets}
