% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{americas_crs}
\alias{americas_crs}
\title{Custom coordinate reference system for American BirdFlow Models}
\format{
\subsection{\code{americas_crs}}{

A \code{crs} object as produced by \code{\link[terra:crs]{terra::crs()}} this is Custom Lamber Azimuthal
projection with an origin of 30, -95 (latitude, longitude),
a false easting of  5.5e6, and a false northing of 9.5e6.
It is designed to show all of North and South America but with
the origin closer to North America so that the distortion is minimal there.
This is an equal area projection but does distort distances and shapes.
}
}
\usage{
americas_crs
}
\description{
Custom coordinate reference system for American BirdFlow Models
}
\keyword{datasets}
