% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banding_csv_collapse_list}
\alias{banding_csv_collapse_list}
\title{Groupings for collapsing raw banding CSV files together before preprocessing.}
\format{
\subsection{\code{banding_csv_collapse_list}}{

A list of length 54, each list item is numeric vector of of CSV file numbers to be processed together
}
}
\source{
\url{https://dx.doi.org/10.5066/P9BSM38F}
}
\usage{
banding_csv_collapse_list
}
\description{
For example, if multiple subspecies of one species are stored in separate files
}
\keyword{datasets}
