% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_availability.R
\name{banding_data_availability}
\alias{banding_data_availability}
\title{Batch data availability info function}
\usage{
banding_data_availability(file)
}
\arguments{
\item{file}{Path to RDS file to check data availability}
}
\value{
data.frame of attrition info
}
\description{
Batch data availability info function
}
