% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_data_banding.R
\name{banding_data_to_linestring}
\alias{banding_data_to_linestring}
\title{Function to convert banding data to linestring of origin-destination pairs}
\usage{
banding_data_to_linestring(rds_file, ...)
}
\arguments{
\item{rds_file}{path to rds file to make linestrings from}

\item{...}{additional arguments passed to make_tracks}
}
\value{
sf data.frame with linestring geometry
}
\description{
Function to convert banding data to linestring of origin-destination pairs
}
