% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banding_raw_file_urls}
\alias{banding_raw_file_urls}
\title{URLs for downloading raw USGS banding datafiles version 2022-07-14}
\format{
\subsection{\code{banding_raw_file_urls}}{

A data frame with 60 obs. of 2 variables, scraped from the html source at the below DOI
\describe{
\item{url}{download URL}
\item{name}{filename}

}
}
}
\source{
\url{https://dx.doi.org/10.5066/P9BSM38F}
}
\usage{
banding_raw_file_urls
}
\description{
URLs for downloading raw USGS banding datafiles version 2022-07-14
}
\keyword{datasets}
