% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_availability.R
\name{batch_data_availability}
\alias{batch_data_availability}
\title{Batch data availability function}
\usage{
batch_data_availability()
}
\value{
Data.frame of data availability
}
\description{
Batch data availability function
}
