% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{batch_evaluate_models}
\alias{batch_evaluate_models}
\title{Batch evaluate models on the cluster}
\usage{
batch_evaluate_models(params, track_info)
}
\arguments{
\item{params}{A list of parameters, as returned by \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}.}

\item{track_info}{A list of track info, as returned by \code{\link[=make_tracks]{make_tracks()}}, and passed internally to \code{\link[BirdFlowR:interval_log_likelihood]{BirdFlowR::interval_log_likelihood()}}.}
}
\value{
A data.frame with a row for each model evaluated
}
\description{
Batch evaluate models on the cluster
}
