% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_flow.R
\name{batch_flow}
\alias{batch_flow}
\title{Grid search, model selection, and model evaluation for one species on the cluster}
\usage{
batch_flow(one_species)
}
\arguments{
\item{one_species}{a character vector of length 1 that is a valid input to \code{\link[ebirdst:get_species]{ebirdst::get_species()}}, e.g., an eBird 6-letter code or a valid common name}
}
\value{
function is used for its many side effects, according to configuration settings in \code{as.list(banding:::the)}
\itemize{
\item create destination directories as needed
\item write preprocessed hdf5 file
\item write modelfit hdf5 files
\item write output files, plots, and maps for model evaluation and visualization
}
}
\description{
Grid search, model selection, and model evaluation for one species on the cluster
}
\seealso{
\code{\link[=multiple_species_batch]{multiple_species_batch()}}
}
