% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_flux.R
\name{batch_flux}
\alias{batch_flux}
\title{Calculate flux for multiple species}
\usage{
batch_flux(
  model_paths,
  flux_paths = NULL,
  base_path = NULL,
  memory = 12,
  walltime = 180,
  show_progress = TRUE
)
}
\arguments{
\item{model_paths}{The path to one or more BirdFlow models saved either
as HDF5 or \link[base:readRDS]{RDS} files.}

\item{flux_paths}{Paths (ending in \code{".rds"}) to save flux to.
It should be the same length as \code{model_paths} or \code{NULL} in which case
\code{"_flux.rds"} will be appended to \code{model_paths} after removing the extension.}

\item{base_path}{The directory in which the batch tools registry directory
should be created (used to manage the run).  If \code{NULL} the directory of
the first path in \code{model_paths} will be used.  Either way the registry
directory will be a concatenation of \code{"batch_flux"}, the date, and \code{"_mf"}.}

\item{memory}{GB of memory allocated for each flux calculation.}

\item{walltime}{Elapsed time in minutes allocated to the job.}

\item{show_progress}{If \code{TRUE} \code{batch_flux()} will display a progress bar
and wait for all the launched jobs to finish. If \code{FALSE} \code{batch_flux()}
will exit immediately after launching.}
}
\value{
Nothing is returned.
}
\description{
Batch calculate flux in parallel using slurm
}
