% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{batch_modelfit_wrapper}
\alias{batch_modelfit_wrapper}
\title{Function to batch fit models on the cluster, including resubmitting any failed jobs}
\usage{
batch_modelfit_wrapper(params)
}
\arguments{
\item{params}{A list of parameters, as generated by \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}}
}
\value{
This function is run for its side-effects, namely calling \code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}} and \code{\link[=birdflow_modelfit]{birdflow_modelfit()}}.
}
\description{
Function to batch fit models on the cluster, including resubmitting any failed jobs
}
\seealso{
\code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}} and \code{\link[=birdflow_modelfit]{birdflow_modelfit()}}
}
