% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banding_data.R
\name{batch_preprocess_raw_files_to_rds}
\alias{batch_preprocess_raw_files_to_rds}
\title{Batch preprocess banding data on the cluster}
\usage{
batch_preprocess_raw_files_to_rds(banding_data_dir)
}
\arguments{
\item{banding_data_dir}{Directory containing CSV files}
}
\value{
side effect, lots of rds files
}
\description{
Batch preprocess banding data on the cluster
}
