% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_species.R
\name{batch_species}
\alias{batch_species}
\title{Fit many species with identical parameters}
\usage{
batch_species(
  species,
  grid_search_type = "old",
  de_ratio = NA,
  obs_prop = NA,
  dist_pow = 0.4167,
  dist_weight = 0.008177,
  ent_weight = 0.001924,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{species}{A vector of species codes or common names to fit. Must work
with \code{\link[ebirdst:get_species]{ebirdst::get_species()}}}

\item{grid_search_type}{Either \code{"old"} to specify all the Python
hyperparameters (\code{dist_pow}, \code{dist_weight}, and
\code{ent_weight}) directly; or\code{"new"} to uses the alternative paramerization of
\code{dist_pow}, \code{obs_prop}, and \code{de_ratio} which are passed to the
\code{refactor_hyperparameters} function to generate the Python parameters.
In either case there is a hidden python parameter \code{obs_weight} which is
always set to 1 when called from \strong{BirdFlowPipeline}.
\code{"new"} is the current default behavior because the ratio between distance
weight and entropy weight heavily determines the behavior of the model.
These hyperparameters are highly sensitive and refactoring them in this
way makes it easier to intuit about and appears to create better
coverage and performance of the "grid" search.}

\item{de_ratio, obs_prop, dist_pow, dist_weight, ent_weight}{loss function
parameters see definition in \code{grid_search_list} below under \code{...}}

\item{show_progress}{If \code{TRUE} \code{batch_species()} will display a progress bar
and wait for all the launched jobs to finish. If \code{FALSE} \code{batch_species()}
will exit immediately after launching.}

\item{...}{
  Arguments passed on to \code{\link[=set_pipeline_params]{set_pipeline_params}}
  \describe{
    \item{\code{gpu_ram}}{Set the GPU RAM allocation in GB when fitting
the models.  It is also passed to \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}} and
if \code{res} is set to \code{NULL} that function will set the \code{res} to the smallest
round number that can be fit with the allocated ram.}
    \item{\code{res}}{The resolution of model. Specifically the height and width of
the model cells in km.  Set to \code{NULL} to fit the finest resolution possible
with the allocated \code{gpu_ram}.}
    \item{\code{suffix}}{Characters to append to the output directory and model file
names.}
    \item{\code{grid_search_list}}{These parameters are  used
in  the loss function by \href{https://github.com/birdflow-science/BirdFlowPy}{BirdFlowPy}.
They are also used in the grid search when running \code{\link[=batch_flow]{batch_flow()}}.
All six possible parameters are in the list even though two of them won't
be used  - which two depends on  \code{grid_search_type}. They are
the set of values that will be combined factorially in the grid
search. The default values represent our currently prefered approach.

\emph{New} and \emph{Old} below indicate which setting of \code{grid_search_type} they
are used with.
\describe{
\item{\code{de_ratio}}{ \emph{New} The ratio of the distance weight to entropy weight
in the loss function that is optimized while fitting the models}
\item{\code{obs_prop}}{ \emph{New} The proportion of the total weight (among
observation, distance, and entropy weights), that is assigned to the
observation weight in the loss function.}
\item{\code{dist_pow}}{ \emph{New and Old} The power used to transform movement
distances prior to weighting them in the loss function}
\item{\code{dist_weight}}{\emph{Old} The weight assigned to distances in the loss
function.}
\item{\code{ent_weight}}{\emph{Old} The weight assigned to entropy in the loss
function.}
}}
    \item{\code{hdf_path}}{The path to the base directory in which hdf5 preprocessed and
fitted models are stored.}
    \item{\code{hdf_dir}}{The path to the directory in which this run's hdf5
preprocessed and fit models are stored. Set by
\code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}} to \verb{<hdf_path>/<species>_<res>km/}.}
    \item{\code{base_output_path}}{The base path for model stats, reports,
and other output derived from the models.
\code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}} to}
    \item{\code{output_fullname}}{Initially \code{NULL} this is set by
\link{preprocess_species_wrapper} to \code{"<species>_<res>km<suffix>"}.}
    \item{\code{output_path}}{Initially \code{NULL} this is set by
\link{preprocess_species_wrapper} to \code{"<base_output_path>/<output_fullname>"}.
Note \code{output_path} is used to store model stats and reports but the model
itself is stored in \code{hdf_dir}.}
    \item{\code{season}}{This is used in two ways:  (1) to filter movement data
(I think just tracks? - ebp) before evaluating the model and (2)
if \code{truncate_season} is \code{TRUE} it is passed to
\code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}}, producing a model that is truncated to
just that season.}
    \item{\code{truncate_season}}{If \code{TRUE} the model will be truncated to \code{season} and
marginals for transitions outside of the season won't be fit or included.}
    \item{\code{model_selection}}{Set how model selection is performed within
\code{\link[=rank_models]{rank_models()}} as called from \code{\link[=batch_flow]{batch_flow()}}.  It should be one of:
\describe{
\item{\code{"str_etc"}}{Straightness and traverse correlation only.}
\item{\code{"pit_etc"}}{PIT metrics and traverse correlation only.}
\item{\code{"real_tracking"}}{Tracking-focused model selection, this
includes traverse correlation; PIT scores; and straightness and
n_stopovers targeted to observed values from real
tracking data.}
\item{\code{"real_tracking_no_cal"}}{Tracking-focused model selection,
this includes traverse correlation; and straightness and n_stopovers
targeted to observed values from real tracking, but no PIT scores.)}
\item{\code{"averaged_parameters"}}{This is a place holder for when the loss
function parameters are fixed - typically to average values - in which
case no model selection is performed.}
}}
    \item{\code{clip}}{This is passed to \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}} to define a
clipping polygon to use while preprocessing - only areas within the polygon
are included in the model.}
    \item{\code{crs}}{Passed to \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}} to define the coordinate
reference system for the model. With the default of \code{NULL} the CRS
is set to the CRS assigned to the species by eBird status and trends.}
    \item{\code{skip_quality_checks}}{Passed to \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}} if \code{TRUE} an
error will be thrown if season quality values in \link[ebirdst:ebirdst_runs]{ebirdst::ebirdst_runs} are
below three.  If \code{TRUE} then attempt to fit the model regardless of quality.}
    \item{\code{fit_only}}{Set to \code{TRUE} to fit a model without a grid search,
model evaluation, model ranking, or model reports.}
    \item{\code{ebirdst_year}}{The version year of the ebirdst package. This shouldn't
be set by users and any supplied value will be ignored.}
    \item{\code{trim_quantile}}{Passed to \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}}.}
  }}
}
\value{
Nothing is returned.
}
\description{
Batch fit many species with fixed loss function parameters without
tuning, evaluation, or reports.
}
