% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_evaluator.R
\name{BatchBirdFlowEvaluator}
\alias{BatchBirdFlowEvaluator}
\title{Construct a batch evaluator}
\usage{
BatchBirdFlowEvaluator(trainer)
}
\arguments{
\item{trainer}{A \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}.}
}
\value{
A \code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}} object (invisible).
}
\description{
Validates a \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}} and returns a corresponding
\code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}} ready to evaluate all fitted HDF5 models for that species.
}
\examples{
\dontrun{
trainer <- BatchBirdFlowTrainer("amewoo", res = 150)
trainer <- fit(trainer)
ev <- BatchBirdFlowEvaluator(trainer)
}
}
\seealso{
\code{\link[=new_BatchBirdFlowEvaluator]{new_BatchBirdFlowEvaluator()}}, \code{\link[=evaluate.BatchBirdFlowEvaluator]{evaluate.BatchBirdFlowEvaluator()}},
\code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}, \code{\link[=fit.BatchBirdFlowTrainer]{fit.BatchBirdFlowTrainer()}}
}
