% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{birdflow_modelfit}
\alias{birdflow_modelfit}
\title{Function to fit BirdFlow model by executing \code{update_hdf.py} in Python.
Arguments typically passed from \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}} via \code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}}}
\usage{
birdflow_modelfit(
  mypy = file.path(the$python_repo_path, "update_hdf.py"),
  mydir,
  mysp,
  myres,
  dist_weight,
  ent_weight,
  dist_pow,
  obs_weight = 1,
  learning_rate = 0.1,
  training_steps = 600,
  rng_seed = 17
)
}
\arguments{
\item{mypy}{Path of \code{update_hdf.py} Python script.}

\item{mydir}{Argument for \code{update_hdf.py}.}

\item{mysp}{Argument for \code{update_hdf.py}.}

\item{myres}{Argument for \code{update_hdf.py}.}

\item{dist_weight}{Argument for \code{update_hdf.py}.}

\item{ent_weight}{Argument for \code{update_hdf.py}.}

\item{dist_pow}{Argument for \code{update_hdf.py}.}

\item{obs_weight}{Argument for \code{update_hdf.py}.}

\item{learning_rate}{Argument for \code{update_hdf.py}.}

\item{training_steps}{Argument for \code{update_hdf.py}.}

\item{rng_seed}{Argument for \code{update_hdf.py}.}
}
\description{
Function to fit BirdFlow model by executing \code{update_hdf.py} in Python.
Arguments typically passed from \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}} via \code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}}
}
\seealso{
\code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}}, \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}
}
