% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{birdflow_modelfit_args_df}
\alias{birdflow_modelfit_args_df}
\title{Create a grid-expanded data.frame of model fit arguments. Output designed to be passed to \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}}
\usage{
birdflow_modelfit_args_df(
  grid_search_type = NULL,
  grid_search_list = NULL,
  hdf_dir = NULL,
  my_species = NULL,
  my_res = NULL
)
}
\arguments{
\item{grid_search_type}{Character, either \code{old} (normal grid) or \code{new} (refactored grid based on \code{\link[=refactor_hyperparams]{refactor_hyperparams()}})}

\item{grid_search_list}{List of arguments and vectors, some of which to \code{grid.expand} over when creating the data.frame:
\itemize{
\item example \code{old} grid search list:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{grid_search_list <- list(
 dist_weight = seq(from = 0.0008, to = 0.0018, length.out = 5),
 ent_weight = seq(from = 0.00015, to = 0.0004, length.out = 5),
 dist_pow = seq(from = 0.1, to = .9, length.out = 5)
)
}\if{html}{\out{</div>}}
\itemize{
\item example \code{new} grid search list:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{grid_search_list <- list(
 de_ratio = c(2, 4, 8, 16),
 obs_prop = c(0.95, 0.975, 0.99, 0.999, 0.9999),
 dist_pow = seq(from = 0.2, to = 0.8, by = 0.15),
 dist_weight = NA_real_,
 ent_weight = NA_real_
)
}\if{html}{\out{</div>}}}

\item{hdf_dir}{Directory containing the prepocessed hdf5 file, and to which modelfit files will be written}

\item{my_species}{Character, species to modelfit as output by \code{\link[ebirdst:get_species]{ebirdst::get_species()}}}

\item{my_res}{Integer, resolution in kilometers}
}
\description{
Create a grid-expanded data.frame of model fit arguments. Output designed to be passed to \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}
}
\seealso{
\code{\link[=birdflow_modelfit]{birdflow_modelfit()}}, \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}
}
