% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_data_banding.R
\name{combine_together_list}
\alias{combine_together_list}
\title{Make the list of file sets required by \code{\link[=process_file_set]{process_file_set()}}}
\usage{
combine_together_list(suffix_vec, together_list)
}
\arguments{
\item{suffix_vec}{A vector of all possible suffixes in the BBL CSV file download set}

\item{together_list}{A list, each element of which is a vector of 2 or more CSV file suffixes that need to be analyzed together}
}
\value{
A list, each element of which is a vector of CSV file suffixes to analyze at the same time. When one of the vectors has length > 1, it's because one or more ebirdst species spans multiple CSV files
}
\description{
Make the list of file sets required by \code{\link[=process_file_set]{process_file_set()}}
}
\seealso{
\code{\link[=process_file_set]{process_file_set()}}
}
