% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banding_data.R
\name{download_banding_files}
\alias{download_banding_files}
\title{Download raw USGS banding files}
\usage{
download_banding_files(banding_raw_path)
}
\arguments{
\item{banding_raw_path}{Path to which to download the files. If not provided, it will download to \code{the$banding_raw_path}}
}
\value{
Function exists for its side effects, namely downloading the files listed in package data object \link{banding_raw_file_urls}
}
\description{
Download raw USGS banding files
}
\seealso{
\link{banding_raw_file_urls}
}
