% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluate_performance_route}
\alias{evaluate_performance_route}
\title{Evaluate how well a BirdFlow model correlates with eBird Status and Trends}
\usage{
evaluate_performance_route(x, season = "all")
}
\arguments{
\item{x}{a BirdFlow object, e.g., the object returned from
\code{BirdFlowR::import_birdflow()}}

\item{season}{a BirdFlow season, e.g., one of the options in
\code{??BirdFlowR::lookup_season_timesteps()}}
}
\value{
a list containing the following:
\itemize{
\item \code{start_cor} correlation between eBird Status and Trends distribution at the
beginning of \code{season} and the model's weekly marginal distribution for
the same week
\item \code{end_traverse_cor} correlation between eBird Status and Trends distribution
at the end of \code{season} and the forward-predicted weekly marginal
distribution for that same week, when starting the prediction from the
model's marginal distribution at \code{season}'s starting week
}
}
\description{
Evaluate how well a BirdFlow model correlates with eBird Status and Trends
}
