% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metrics.R
\name{format_metrics}
\alias{format_metrics}
\title{Format metrics for plotting}
\usage{
format_metrics(
  train_metrics,
  test_metrics,
  cor_threshold,
  cor_metric,
  this_model = NULL,
  include = NULL,
  use = "test",
  long = TRUE
)
}
\arguments{
\item{train_metrics, test_metrics}{Similarly formatted evaluation metrics
for all the models in the  grid search.  This is saved to an Rds during Yangkang's grid search
workflow and is calculated on either the training models or testing models.}

\item{cor_threshold}{A threshold in correlation between model distributions
and training (S&T) distributions below which models will be excluded
for selection by some of the criteria.}

\item{cor_metric}{The particular correlation metric to which the
threshold is applied -  a column in metrics.}

\item{this_model}{The model id (in \code{metrics$model}) for the current model.
This is used when making a model report to indicate which model the report
is for,}

\item{include}{list of metric columns to include in the output.}

\item{use}{either \code{"test"} or \code{"train"} indicating which metrics should be
plotted.}

\item{long}{if \code{TRUE} use long format with metric and value column,
otherwise use wider format with a column for each metric.}
}
\value{
A modified version of \code{metrics} with
models labeled by type. Will include both the best and candidate models.
Columns:
\item{model}{the model file name - used as unique ID}
\item{type}{Classified model types. One of the types defined by \code{\link[=identify_best_models]{identify_best_models()}} or one of
\code{this}, \code{candidate} or \code{low_cor}}
\item{metric}{one of the metric names - a column in the input}
\item{value}{the metric value.}
}
\description{
This formats the metris
}
