% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transitions.R
\name{get_ground_truth_routes_intervals_and_one_week_intervals}
\alias{get_ground_truth_routes_intervals_and_one_week_intervals}
\title{Functions to get ground truth routes and transitions for model tuning and validation, combining tracking, Motus, and banding.}
\usage{
get_ground_truth_routes_intervals_and_one_week_intervals(params, bf)
}
\arguments{
\item{params}{The params list preprocessed by \code{set_pipeline_params} and \code{preprocess_species_wrapper} functions.}

\item{bf}{The BirdFlow model, used to snap routes into BirdFlow coordinates.}
}
\value{
combined_routes_data, interval_obj, interval_one_week_obj
}
\description{
Functions to get ground truth routes and transitions for model tuning and validation, combining tracking, Motus, and banding.
}
