% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_data.R
\name{get_real_track}
\alias{get_real_track}
\title{Get the BirdFlowRoutes object of the real track}
\usage{
get_real_track(bf, params, filter = FALSE)
}
\arguments{
\item{bf}{The BirdFlow model object}

\item{params}{The params object, containing at least the hdf_dir from which to learn bf object}
}
\value{
the BirdFlowRoutes object of the real track
}
\description{
Get the BirdFlowRoutes object of the real track
}
