% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_ram.R
\name{gpu_ram}
\alias{gpu_ram}
\title{Calculate the GPU RAM needed for a model}
\usage{
gpu_ram(bf, max_param_per_gpu_gb = NULL)
}
\arguments{
\item{bf}{A BirdFlow model object}

\item{max_param_per_gpu_gb}{The number of parameters that can be fit with
each GB of GPU ram. If \code{NULL}, the default the value from
\code{\link[BirdFlowR:birdflow_options]{BirdFlowR::birdflow_options()}}` is used.
The only reason to set it here is if that value is empirically wrong.}
}
\value{
The GB of GPU RAM that was or will be required to fit \code{bf}
}
\description{
Calculate the GPU RAM needed for a model
}
\examples{
bf <- BirdFlowModels::amewoo
gpu_ram(bf)
}
