% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_best_model.R
\name{identify_best_models}
\alias{identify_best_models}
\title{Identify the best models in grid search}
\usage{
identify_best_models(metrics, cor_threshold = 0.9, cor_metric = "min_dist_cor")
}
\arguments{
\item{metrics}{A table of evaluation metrics for all the models in the
grid search.  This is saved to an Rds during Yangkang's grid search
workflow and can be calculated on either the training models
(often \code{train}) or the reserved testing models (\code{test}).}

\item{cor_threshold}{A threshold in correlation between model distributions
and training (S&T) distributions below which models will be excluded
for selection by some of the criteria.}

\item{cor_metric}{The particular correlation metric to which the
threshold is applied -  a column in metrics.}
}
\value{
A data.frame with:
\item{type}{the selection criteria used to select the model in its short form}
\item{model}{the model id as it apears in \code{metrics$model}}
}
\description{
Identify the best models from training or test metrics calculated for all
the models in a grid search.
\code{identify_best_models()} duplicates Yangkang's model selection code
and so needs to be updated as he changes
the selection criteria.
}
\details{
Note the appropriate value for \code{cor_threshold} depends on model resolution.
To get an equivalent amount of filtering a lower threshold is needed
when the cell size is smaller.

Last update 2025-12-12
}
