% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_trainer.R
\name{identify_hdf5_model}
\alias{identify_hdf5_model}
\title{Read key parameters from hdf5 modelfit}
\usage{
identify_hdf5_model(hdf5_path)
}
\arguments{
\item{hdf5_path}{Path of hdf5 model to check}
}
\value{
a one-row data.frame of hdf5 info
}
\description{
Read key parameters from hdf5 modelfit
}
