% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_birdflow_and_evaluate}
\alias{import_birdflow_and_evaluate}
\title{Import BirdFlow model, and apply further arguments to \code{\link[=evaluate_model]{evaluate_model()}}}
\usage{
import_birdflow_and_evaluate(path, ...)
}
\arguments{
\item{path}{Path of hdf5 to pass to \code{\link[BirdFlowR:import_birdflow]{BirdFlowR::import_birdflow()}}.}

\item{...}{Further arguments passed to \code{\link[=evaluate_model]{evaluate_model()}}}
}
\description{
Import BirdFlow model, and apply further arguments to \code{\link[=evaluate_model]{evaluate_model()}}
}
