% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_banding_data}
\alias{load_banding_data}
\title{Load banding data}
\usage{
load_banding_data(
  banding_rds_path,
  max_band_tracks = 10000,
  max_preprocess = 5e+05,
  min_dist_m = 15000,
  max_days = 180
)
}
\arguments{
\item{banding_rds_path}{file path to \code{.rds} file containing raw banding data}

\item{max_band_tracks}{maximum number of band_tracks to output, downsample if necessary}

\item{max_preprocess}{maximum number of bands to preprocess, downsample if necessary}

\item{min_dist_m}{minimum distance in meters between origin and destination to retain a band_track}

\item{max_days}{maximum number of days between origin and destination to retain a band_track}
}
\value{
a list:
\itemize{
\item \code{banding_df} data.frame of observations for banding data
}
}
\description{
Load banding data
}
