% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_data_banding.R
\name{load_banding_df}
\alias{load_banding_df}
\title{Load banding data}
\usage{
load_banding_df(banding_rds_path)
}
\arguments{
\item{banding_rds_path}{The banding RDS path of the species}
}
\description{
Load banding data
}
