% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ground_truth_transition_pairs.R
\name{load_banding_transitions_df}
\alias{load_banding_transitions_df}
\title{Load banding data}
\usage{
load_banding_transitions_df(banding_rds_path)
}
\arguments{
\item{banding_rds_path}{The banding RDS path of the species}
}
\description{
Load banding data
}
