% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ground_truth_transition_pairs.R
\name{load_motus_transitions_df}
\alias{load_motus_transitions_df}
\title{Load Motus data}
\usage{
load_motus_transitions_df(motus_rds_path)
}
\arguments{
\item{motus_rds_path}{The motus RDS path of the species}
}
\description{
Load Motus data
}
