% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_3d_plot}
\alias{make_3d_plot}
\title{3d plot function}
\usage{
make_3d_plot(color_column, suffix, eval_metrics, params)
}
\arguments{
\item{color_column}{column to use for the colors}

\item{suffix}{suffix to use for output filenames}

\item{eval_metrics}{data.frame produced from rbindlist-ing output from \code{\link[=evaluate.BatchBirdFlowEvaluator]{evaluate.BatchBirdFlowEvaluator()}}}

\item{params}{the standard params list object, see \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}}
}
\value{
side-effect is a plot file created
}
\description{
3d plot function
}
\seealso{
\code{\link[=evaluate.BatchBirdFlowEvaluator]{evaluate.BatchBirdFlowEvaluator()}}, \code{\link[=evaluate_model]{evaluate_model()}}, \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}
}
