% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flux_index.R
\name{make_flux_index}
\alias{make_flux_index}
\title{Make html index file for flux reports}
\usage{
make_flux_index(index, html)
}
\arguments{
\item{index}{Data frame with columns:
\describe{
\item{\code{species}}{eBird species code}
\item{\code{common_name}}{Species common name}
\item{\code{scientific}}{Scientific name}
\item{\code{report_exists}}{TRUE if an html report exists for the species.}
\item{\code{rel_link}}{The relative link to the species report from the index location (\code{html})}
}}

\item{html}{path to output html index file.}
}
\value{
Nothing is returned, a report is created at \code{html}
}
\description{
Note: I ran this on unity after using \code{devtools::load_all()} it
seemed to be writing temporary files to the inst/rmd directory.  This
suggests that it might not work with an installed version of BirdFlowPipeline.
If that's the case we'd have to copy the Rmd from the package to a temorary
location before knitting.
}
\examples{
\dontrun{
index <- readRDS(system.file("rmd/flux_index_example.Rds", package = "BirdFlowPipeline"))
html <- file.path(tempdir(), "index.html")
make_flux_index(index, html)
file.exists(html)
file.remove(html)
}
}
