% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flux_report.R
\name{make_flux_report}
\alias{make_flux_report}
\title{Make flux report}
\usage{
make_flux_report(hdf, flux, anim, html, prevlink, nextlink)
}
\arguments{
\item{hdf}{path to HDF5 model file}

\item{flux}{path to flux RDS file}

\item{anim}{path to abundance and flux animation GIF}

\item{html}{path to output html file}

\item{prevlink}{relative path to the previous species report or NULL if
there is none.}

\item{nextlink}{relative path to the next species report or NULL if none.}
}
\value{
Nothing is returned, a report is created at \code{html}
}
\description{
Make flux report
}
