% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_report.R
\name{make_new_report}
\alias{make_new_report}
\title{Make species report}
\usage{
make_new_report(
  model,
  bmtr,
  anim,
  html,
  routes,
  prevlink = NA,
  nextlink = NA,
  test_metrics = NA,
  train_metrics = NA,
  title = NA,
  cor_threshold = NA,
  cor_metric = NA,
  mean_rating = NA
)
}
\arguments{
\item{model}{path to HDF5 model file}

\item{bmtr}{path to BMTR RDS file containing output from \code{\link[BirdFlowR:calc_bmtr]{BirdFlowR::calc_bmtr()}}}

\item{anim}{path to abundance and flux animation GIF}

\item{html}{path to output html file}

\item{routes}{path to routes plot}

\item{prevlink}{relative path to the previous species report or NULL if
there is none.}

\item{nextlink}{relative path to the next species report or NULL if none.}

\item{test_metrics, train_metrics}{Optional data frame containing
performance metrics from the grid search for the training and test
datasets respectively. If NA than the tuning
section isn't included in the report.
@param title Title for report, default of \code{NA} results in  "\if{html}{\out{<Common name>}} BMTR"
@param cor_threshold,cor_metric Threshold and metric used to exclude
models that are not sufficient close to the training S&T distributions}

\item{mean_rating}{Mean rating from 1 (worst) to 5 (best),
assigned to the BirdFlow model
by a minimum of two reviewers on the BirdFlow team.}
}
\value{
Nothing is returned, a report is created at \code{html}
}
\description{
Note: I ran this on unity after using \code{devtools::load_all()} it
seemed to be writing temporary files to the inst/rmd directory.  This
suggests that it might not work with an installed version of BirdFlowPipeline.
If that's the case we'd have to copy the Rmd from the package to a temorary
location before knitting.
}
\examples{
\dontrun{

p <- list()
p$model <-  system.file("rmd/example_data/sstspa_model.hdf5", package = "BirdFlowPipeline") 
p$bmtr <-  system.file("rmd/example_data/sstspa_bmtr.Rds", package = "BirdFlowPipeline") 
p$anim <-  system.file("rmd/example_data/sstspa_anim.gif", package = "BirdFlowPipeline") 
p$html <-  tempfile(fileext = ".html", pattern = "report")
p$routes <-  system.file("rmd/example_data/sstspa_routes.hdf5", package = "BirdFlowPipeline") 
p$prevlink <-  NA
p$nextlink <-  NA
p$test_metrics <-  system.file("rmd/example_data/sstspa_test_metrics.rds", package = "BirdFlowPipeline") |> readRDS()                             
p$train_metrics <-  system.file("rmd/example_data/sstspa_train_metrics.rds", package = "BirdFlowPipeline")   |> readRDS()
p$title <- "Example Species Report"
p$cor_threshold <- 0.9
p$cor_metric <- "min_dist_cor"  
p$mean_rating <- 4

do.call(make_new_report, args = p)

}
}
