% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_species_index.R
\name{make_species_index}
\alias{make_species_index}
\title{Function to make an index file for output of batch_species()}
\usage{
make_species_index(hdf_path, res)
}
\arguments{
\item{hdf_path}{Path to the directory}

\item{res}{Model resolution}
}
\value{
It invisibly returns the index
}
\description{
Searches through a directory containing output from \code{batch_species()} and
create an index file with  columns:
"species", "hdf5", "size", "breeding_quality", "nonbreeding_quality",
"postbreeding_migration_quality", "prebreeding_migration_quality",
"min_quality", "xmin", "xmax", "ymin", "ymax")
}
