% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{model_evaluation_biplot}
\alias{model_evaluation_biplot}
\title{PCA biplot hyperparameters evaluation}
\usage{
model_evaluation_biplot(ll_df, params)
}
\arguments{
\item{ll_df}{data.frame produced from rbindlist-ing output from \code{\link[=batch_evaluate_models]{batch_evaluate_models()}}}

\item{params}{the standard params list object, see \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}}
}
\value{
side effect is a plot written to file
}
\description{
PCA biplot hyperparameters evaluation
}
\seealso{
\code{\link[=batch_evaluate_models]{batch_evaluate_models()}}, \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}
}
