% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_flow.R
\name{multiple_species_batch}
\alias{multiple_species_batch}
\title{Grid search, model selection, and model evaluation for multiple species on the cluster, with error handling}
\usage{
multiple_species_batch(multispecies_vector)
}
\arguments{
\item{multispecies_vector}{a character vector that contains a valid inputs to \code{\link[ebirdst:get_species]{ebirdst::get_species()}}, e.g., eBird 6-letter codes or valid common names.}
}
\value{
function is used for its many side effects, according to configuration settings in \code{as.list(banding:::the)}
\itemize{
\item create destination directories as needed
\item write preprocessed hdf5 files
\item write modelfit hdf5 files
\item write output files, plots, and maps for model evaluation and visualization
}
}
\description{
Grid search, model selection, and model evaluation for multiple species on the cluster, with error handling
}
\seealso{
\code{\link[=batch_flow]{batch_flow()}} for doing this for one species
}
