% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_evaluator.R
\name{new_BatchBirdFlowEvaluator}
\alias{new_BatchBirdFlowEvaluator}
\title{Initialize a batch evaluator (low-level constructor)}
\usage{
new_BatchBirdFlowEvaluator(trainer)
}
\arguments{
\item{trainer}{A \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}.}
}
\value{
A \code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}} with field \code{batch_trainer}.
}
\description{
Creates a \code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}} that carries a reference to the
underlying \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}.
}
\examples{
\dontrun{
ev <- new_BatchBirdFlowEvaluator(BatchBirdFlowTrainer("amewoo"))
}
}
\seealso{
\code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}}, \code{\link[=evaluate.BatchBirdFlowEvaluator]{evaluate.BatchBirdFlowEvaluator()}}
}
