% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_loader.R
\name{new_transitions_loader}
\alias{new_transitions_loader}
\title{Initialize a transitions loader (low-level constructor)}
\usage{
new_transitions_loader(batch_trainer)
}
\arguments{
\item{batch_trainer}{A \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}.}
}
\value{
A \code{\link[=TransitionsLoader]{TransitionsLoader()}} with field \code{batch_trainer}.
}
\description{
Builds a \code{\link[=TransitionsLoader]{TransitionsLoader()}} that carries a reference to the associated
\code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}.
}
\examples{
\dontrun{
tl <- new_transitions_loader(trainer)
}
}
\seealso{
\code{\link[=TransitionsLoader]{TransitionsLoader()}}, \code{\link[=load.TransitionsLoader]{load.TransitionsLoader()}}
}
