% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gridsearch_metrics.R
\name{plot_grid_search_metrics}
\alias{plot_grid_search_metrics}
\title{Plot grid search metrics}
\usage{
plot_grid_search_metrics(
  train_metrics,
  test_metrics,
  bf = NULL,
  include = NULL,
  cor_threshold,
  cor_metric,
  use = "test"
)
}
\arguments{
\item{train_metrics}{training statistics data frame}

\item{test_metrics}{test statistics data frame}

\item{bf}{bird flow model}

\item{include}{Names of metrics to include in the plots should be columns
in \code{train_metrics} or \code{test_metrics}.}

\item{cor_threshold}{correlation threshold used when selecting models}

\item{cor_metric}{correlation metrics used when selecting models}

\item{use}{\code{"test"} or \code{"train"} to indicate what statistics to use in the plot}
}
\value{
A ggplot plot object
}
\description{
Plot the selected model metrics relative to the
distribution of across all models in the grid search
}
\examples{
files <- c("sstspa_model.hdf5", "sstspa_test_metrics.rds", "sstspa_train_metrics.rds")
paths <- lapply(paste0("rmd/example_data/", files), function(x) system.file(x, package = "BirdFlowPipeline"))
names(paths) <- c("model", "test_metrics", "train_metrics")
bf <- import_birdflow(paths$model)
test_metrics <- readRDS(paths$test_metrics)
train_metrics <- readRDS(paths$train_metrics)

included_metrics <- c("mean_dist_cor", "mean_ll", "mean_energy_improvement", 
  "weighted_mean_win_prob", "weighted_mean_win_distance",
  "weighted_mean_win_distance_fraction", "weighted_mean_ll", 
  "weighted_energy_improvement", "pit_row", "pit_col")

plot_grid_search_metrics(train = train_metrics, test = test_metrics, bf, include = included_metrics,  cor_threshold = 0.9, cor_metric = "min_dist_cor")

}
