% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tuning.R
\name{plot_tuning}
\alias{plot_tuning}
\title{Plot gridsearch performance metrics}
\usage{
plot_tuning(
  train,
  test,
  bf,
  cor_threshold,
  cor_metric = "min_dist_cor",
  use = "test"
)
}
\arguments{
\item{train}{training statistics data frame}

\item{test}{test statistics data frame}

\item{bf}{bird flow model}

\item{cor_threshold}{correlation threshold used when selecting models}

\item{cor_metric}{correlation metrics used when selecting models}

\item{use}{\code{"test"} or \code{"train"} to indicate what statistics to use in the plot}
}
\value{
A ggplot2 plot object
}
\description{
Plot gridsearch performance metrics
}
\examples{

files <- c("sstspa_model.hdf5", "sstspa_test_metrics.rds", "sstspa_train_metrics.rds")
paths <- lapply(paste0("rmd/example_data/", files), function(x) system.file(x, package = "BirdFlowPipeline"))
names(paths) <- c("model", "test_metrics", "train_metrics")
bf <- import_birdflow(paths$model)
test_metrics <- readRDS(paths$test_metrics)
train_metrics <- readRDS(paths$train_metrics)

plot_tuning(train = train_metrics, test = test_metrics, bf, 0.9 )

}
