% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_and_fit.R
\name{preprocess_and_fit}
\alias{preprocess_and_fit}
\title{Process and fit a single model with fixed parameters}
\usage{
preprocess_and_fit(species, params)
}
\arguments{
\item{species}{}

\item{params}{A parameter list similar to that returned by
\code{\link[=set_pipleline_paramaters]{set_pipleline_paramaters()}} but for this function each of the
\code{grid_search} parameter list items should be a single value.}
}
\value{
This function is called for the side affect of creating a fitted
hdf5 model file.
}
\description{
This is primarily intended for use by \code{\link[=batch_species]{batch_species()}} but could
also be called directly to fit a single species.
}
