% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{preprocess_species_wrapper}
\alias{preprocess_species_wrapper}
\title{Wrapper for \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}} that also prepares batch parameters}
\usage{
preprocess_species_wrapper(params)
}
\arguments{
\item{params}{A list of starting parameters, currently set up in \code{\link[=batch_flow]{batch_flow()}}}
}
\value{
The starting params list, modified to include additional information calculated during preprocessing. Also has side effects:
\itemize{
\item create directories as needed
\item write preprocessed hdf5 file
}
}
\description{
Wrapper for \code{\link[BirdFlowR:preprocess_species]{BirdFlowR::preprocess_species()}} that also prepares batch parameters
}
\seealso{
\code{\link[=batch_flow]{batch_flow()}}
}
