% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banding_data.R
\name{process_file_set}
\alias{process_file_set}
\title{Function to take a set of raw files and produce per-species RDS}
\usage{
process_file_set(collapse_list_item)
}
\arguments{
\item{collapse_list_item}{A vector containing CSV file numbers to process together}
}
\value{
Side effect is writing \code{.rds} file or files
}
\description{
Function to take a set of raw files and produce per-species RDS
}
