% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_flux.R
\name{process_flux}
\alias{process_flux}
\title{Import a BirdFlow model, calculate flux, and save to an RDS file}
\usage{
process_flux(model_path, flux_path)
}
\arguments{
\item{model_path}{Path to a BirdFlow model file, either an HDF5
to be read with \code{\link[BirdFlowR:export_import_birdflow]{BirdFlowR::import_birdflow()}} or and RDS handled by
\code{\link[=readRDS]{readRDS()}}.}

\item{flux_path}{The path to and RDS file (ending in \code{".rds"}) to write the
result to.}
}
\value{
Flux is written to disk nothing is returned.
}
\description{
\code{process_flux()} is a wrapper to \code{\link[=calc_flux]{calc_flux()}} that handles reading the input
model and writing the resulting flux to disk.
}
