% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quick_visualize_routes}
\alias{quick_visualize_routes}
\title{Quick visualize a spring migration routes simulation from eval_metrics by row number}
\usage{
quick_visualize_routes(
  i,
  n = 10,
  season = "prebreeding",
  df = eval_metrics,
  dir = hdf_dir
)
}
\arguments{
\item{i}{row number in \code{df} from which to extract model}

\item{n}{number of routes to simulate for visualization}

\item{season}{season for which to simulate route, parseable by \code{\link[BirdFlowR:lookup_season_timesteps]{BirdFlowR::lookup_season_timesteps()}}}

\item{df}{a \code{eval_metrics} produced by \code{\link[=evaluate.BatchBirdFlowEvaluator]{evaluate.BatchBirdFlowEvaluator()}}}

\item{dir}{hdf5 directory from \code{params} list.  Need to refactor some of these arguments}
}
\value{
side effect is an interactive plot
}
\description{
Quick visualize a spring migration routes simulation from eval_metrics by row number
}
\seealso{
\code{BirdFlowR::lookup_season_timesteps()}, \code{\link[=evaluate.BatchBirdFlowEvaluator]{evaluate.BatchBirdFlowEvaluator()}}
}
