% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rank_models}
\alias{rank_models}
\title{Rank the models in a ll_df}
\usage{
rank_models(ll_df, params)
}
\arguments{
\item{ll_df}{The ll_df, often produced by recombining results from \code{\link[=batch_evaluate_models]{batch_evaluate_models()}}}

\item{params}{The standard params list, which includes \code{model_selection} method}
}
\value{
Updated ll_df including desirability columns
}
\description{
Rank the models in a ll_df
}
