% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_data.R
\name{real_track_stats}
\alias{real_track_stats}
\title{Calculate rts_stats from real tracks file(s)}
\usage{
real_track_stats(eval_metrics, params)
}
\arguments{
\item{eval_metrics}{The eval_metrics data frame containing at least one valid model name, from which to load bf object for CRS/grid}

\item{params}{The params object, containing at least the hdf_dir from which to learn bf object}
}
\value{
rts_stats from the actual tracks
}
\description{
Calculate rts_stats from real tracks file(s)
}
