% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_release.R
\name{rebuild_model_reports}
\alias{rebuild_model_reports}
\title{Rebuild model report HTML files from existing output directory}
\usage{
rebuild_model_reports(output_directory, model_numbers = 1:5)
}
\arguments{
\item{output_directory}{The output directory}

\item{model_numbers}{Number of top models to rebuild Html files for}
}
\value{
Side effects, rebuilt Html files in output directory
}
\description{
Rebuild model report HTML files from existing output directory
}
