% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_trainer.R
\name{refactor_hyperparams}
\alias{refactor_hyperparams}
\title{Find a dist_weight and ent_weight consistent with a given obs_prop and de_ratio}
\usage{
refactor_hyperparams(de_ratio, obs_prop, digits = 5)
}
\arguments{
\item{de_ratio}{numeric Desired ratio, dist_weight/ent_weight}

\item{obs_prop}{numeric Desired observation_proportion when obs_weight = 1,
e.g. obs_prop = 1 / (1 + dist_weight + ent_weight)}

\item{digits}{integer Desired number of significant digits to retain in result}
}
\value{
A list:
\itemize{
\item \code{dist_weight} dist_weight parameter that satisfies desired de_ratio and obs_prop
\item \code{ent_weight} ent_weight parameter that satisfies desired de_ratio and obs_prop
}
}
\description{
Find a dist_weight and ent_weight consistent with a given obs_prop and de_ratio
}
