% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rts_stats}
\alias{rts_stats}
\title{Calculate route summary statistics}
\usage{
rts_stats(rts)
}
\arguments{
\item{rts}{data.frame object output from \code{BirdFlowR::route()} or \code{BirdFlowR::route_migration()}}
}
\value{
a list of mean summary statistics
}
\description{
Calculate route summary statistics
}
\seealso{
\code{\link[BirdFlowR:route]{BirdFlowR::route()}}, \code{\link[BirdFlowR:route_migration]{BirdFlowR::route_migration()}}
}
