% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{safe_numeric}
\alias{safe_numeric}
\title{Safely return a real value}
\usage{
safe_numeric(x)
}
\arguments{
\item{x}{Input}
}
\value{
x if x is not null, else NA_real_
}
\description{
Safely return a real value
}
