% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_loader.R
\name{split}
\alias{split}
\title{Generic split S3}
\usage{
split(object, ...)
}
\arguments{
\item{object}{The object to split the data for.}

\item{...}{Passed to methods.}
}
\value{
Method-dependent; see specific methods.
}
\description{
Dispatches to class-specific \code{split()} methods (e.g.,
\code{\link[=split.TransitionsLoader]{split.TransitionsLoader()}}).
}
\examples{
\dontrun{
parts <- split(load(transitions_loader(trainer)), seed = 123)
}
}
