% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_release.R
\name{stage_model_release}
\alias{stage_model_release}
\title{Copy model file and html model report from pipeline output to release staging folder}
\usage{
stage_model_release(
  output_directory,
  desirability_rank,
  staging_directory = file.path("/work/pi_drsheldon_umass_edu/birdflow_modeling",
    "dslager_umass_edu/model_release_staging")
)
}
\arguments{
\item{output_directory}{dir where pipeline outputs models and reports}

\item{desirability_rank}{desirability rank of model to release (1 = top model)}

\item{staging_directory}{where to put release staging models and Rds files}
}
\value{
side effects, copies Rds and html files to staging folder
}
\description{
Copy model file and html model report from pipeline output to release staging folder
}
