% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_loader.R
\name{TransitionsLoader}
\alias{TransitionsLoader}
\title{Create a transitions loader}
\usage{
TransitionsLoader(batch_trainer)
}
\arguments{
\item{batch_trainer}{A \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}.}
}
\value{
A \code{\link[=TransitionsLoader]{TransitionsLoader()}} object (invisible).
}
\description{
Validates a \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}} and constructs a \code{\link[=TransitionsLoader]{TransitionsLoader()}}
that can load and split combined movement data for evaluation/tuning.
}
\examples{
\dontrun{
tl <- TransitionsLoader(trainer)
}
}
\seealso{
\code{\link[=new_transitions_loader]{new_transitions_loader()}}, \code{\link[=load.TransitionsLoader]{load.TransitionsLoader()}}, \code{\link[=split.TransitionsLoader]{split.TransitionsLoader()}}
}
