% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_batch_trainer}
\alias{validate_batch_trainer}
\title{What a batch trainer should have:
\enumerate{
\item params
}}
\usage{
validate_batch_trainer(trainer)
}
\description{
What a batch trainer should have:
\enumerate{
\item params
}
}
