% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_classes.R
\name{validate_BatchBirdFlowEvaluator}
\alias{validate_BatchBirdFlowEvaluator}
\title{Validate a BatchBirdFlowEvaluator object}
\usage{
validate_BatchBirdFlowEvaluator(evaluator)
}
\arguments{
\item{evaluator}{An object to validate.}
}
\value{
Invisible \code{NULL}. Throws an error if validation fails.
}
\description{
Ensures that the input is a \code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}} and that it
contains the required attributes.
}
\details{
Checks:
\itemize{
\item Must inherit from both \code{"list"} and \code{"BatchBirdFlowEvaluator"}.
\item Must contain element \code{"batch_trainer"}.
}
}
\examples{
\dontrun{
ev <- BatchBirdFlowEvaluator(batch_trainer("amewoo"))
validate_BatchBirdFlowEvaluator(ev)
}
}
\seealso{
\code{\link[=BatchBirdFlowEvaluator]{BatchBirdFlowEvaluator()}}, \code{\link[=validate_BatchBirdFlowTrainer]{validate_BatchBirdFlowTrainer()}}
}
