% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_classes.R
\name{validate_BatchBirdFlowTrainer}
\alias{validate_BatchBirdFlowTrainer}
\title{Validate a BatchBirdFlowTrainer object}
\usage{
validate_BatchBirdFlowTrainer(trainer)
}
\arguments{
\item{trainer}{An object to validate.}
}
\value{
Invisible \code{NULL}. Throws an error if validation fails.
}
\description{
Ensures that the input is a \code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}} and that it
contains the required attributes.
}
\details{
Checks:
\itemize{
\item Class must include \code{"BatchBirdFlowTrainer"}.
\item Must contain element \code{"params"}.
}
}
\examples{
\dontrun{
validate_batch_trainer(batch_trainer("amewoo"))
}
}
\seealso{
\code{\link[=BatchBirdFlowTrainer]{BatchBirdFlowTrainer()}}, \code{\link[=validate_BatchBirdFlowEvaluator]{validate_BatchBirdFlowEvaluator()}},
\code{\link[=validate_TransitionsLoader]{validate_TransitionsLoader()}}
}
