% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_classes.R
\name{validate_TransitionsLoader}
\alias{validate_TransitionsLoader}
\title{Validate a TransitionsLoader object}
\usage{
validate_TransitionsLoader(loader)
}
\arguments{
\item{loader}{An object to validate.}
}
\value{
Invisible \code{NULL}. Throws an error if validation fails.
}
\description{
Ensures that the input is a \code{\link[=TransitionsLoader]{TransitionsLoader()}} and that it
contains the required attributes.
}
\details{
Checks:
\itemize{
\item Class must include \code{"TransitionsLoader"}.
\item Must contain element \code{"batch_trainer"}.
}
}
\examples{
\dontrun{
tl <- transitions_loader(BatchBirdFlowTrainer("amewoo"))
validate_TransitionsLoader(tl)
}
}
\seealso{
\code{\link[=TransitionsLoader]{TransitionsLoader()}}, \code{\link[=validate_split_data]{validate_split_data()}}
}
