% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{batch_evaluate_models}
\alias{batch_evaluate_models}
\title{Batch evaluate models on the cluster}
\usage{
batch_evaluate_models(params, birdflow_intervals, birdflow_intervals_one_week)
}
\arguments{
\item{params}{A list of parameters, as returned by \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}.}

\item{birdflow_intervals}{A BirdFlowIntervals object.}
}
\value{
A data.frame with a row for each model evaluated
}
\description{
Batch evaluate models on the cluster
}
\seealso{
\code{\link[=evaluate_model]{evaluate_model()}}, \code{\link[=preprocess_species_wrapper]{preprocess_species_wrapper()}}, \code{\link[=as_BirdFlowIntervals]{as_BirdFlowIntervals()}}, \code{\link[BirdFlowR:interval_log_likelihood]{BirdFlowR::interval_log_likelihood()}}
}
