% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_flow.R
\name{batch_flow}
\alias{batch_flow}
\title{Grid search, model selection, and model evaluation for one species on the cluster}
\usage{
batch_flow(
  species,
  params = list(species = character(0), gpu_ram = 10, res = 150, suffix =
    as.character(Sys.Date()), grid_search_type = "new", grid_search_list = list(de_ratio
    = c(2, 4, 6, 8, 10), obs_prop = c(0.95, 0.975, 0.99, 0.999, 0.9999), dist_pow =
    seq(from = 0.1, to = 0.9, by = 0.1), dist_weight = NA_real_, ent_weight = NA_real_),
    hdf_path = the$hdf_path, output_path = the$output_path, season = "prebreeding",
    truncate_season = FALSE, model_selection = "real_tracking", clip = NULL,
    skip_quality_checks = FALSE, fit_only = FALSE)
)
}
\arguments{
\item{species}{a character vector of length 1 that is a valid input to \code{\link[ebirdst:get_species]{ebirdst::get_species()}}, e.g., an eBird 6-letter code or a valid common name}

\item{params}{a list of standard parameters, typically constructed by complex default argument.
\itemize{
\item example \code{old} grid search list:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{grid_search_list <- list(
 dist_weight = seq(from = 0.0008, to = 0.0018, length.out = 5),
 ent_weight = seq(from = 0.00015, to = 0.0004, length.out = 5),
 dist_pow = seq(from = 0.1, to = .9, length.out = 5)
)
}\if{html}{\out{</div>}}
\itemize{
\item example \code{new} grid search list:
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{grid_search_list <- list(
 de_ratio = c(2, 4, 8, 16),
 obs_prop = c(0.95, 0.975, 0.99, 0.999, 0.9999),
 dist_pow = seq(from = 0.2, to = 0.8, by = 0.15),
 dist_weight = NA_real_,
 ent_weight = NA_real_
)
}\if{html}{\out{</div>}}}
}
\value{
function is used for its many side effects, according to configuration settings in \code{as.list(BirdFlowPipeline:::the)}
\itemize{
\item create destination directories as needed
\item write preprocessed hdf5 file
\item write modelfit hdf5 files
\item write output files, plots, and maps for model evaluation and visualization
}
}
\description{
Grid search, model selection, and model evaluation for one species on the cluster
}
\seealso{
\code{\link[=multiple_species_batch]{multiple_species_batch()}}
}
