% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_banding_data.R
\name{batch_preprocess_raw_files_to_rds}
\alias{batch_preprocess_raw_files_to_rds}
\title{Batch preprocess banding data on the cluster}
\usage{
batch_preprocess_raw_files_to_rds(banding_raw_path, banding_rds_path)
}
\arguments{
\item{banding_raw_path}{Custom path with raw banding CSV files}

\item{banding_rds_path}{Custom path to output RDS banding files}
}
\value{
side effect, lots of rds files
}
\description{
Batch preprocess banding data on the cluster
}
