% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_trainer.R
\name{birdflow_modelfit}
\alias{birdflow_modelfit}
\title{Function to fit BirdFlow model by executing \code{update_hdf.py} in Python.
Arguments typically passed from \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}} via
\code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}}}
\usage{
birdflow_modelfit(
  py_script = file.path(the$python_repo_path, "update_hdf.py"),
  dir,
  species,
  res,
  dist_weight,
  ent_weight,
  dist_pow,
  obs_weight = 1,
  learning_rate = 0.1,
  training_steps = 600,
  rng_seed = 17,
  ebirdst_year
)
}
\arguments{
\item{py_script}{Path of Python script to fit a model (\code{update_hdf.py})}

\item{dir}{Argument for \code{update_hdf.py} - the directory for output}

\item{species}{Argument for \code{update_hdf.py} - the species code to fit}

\item{res}{Argument for \code{update_hdf.py} - the model resolution}

\item{dist_weight}{Argument for \code{update_hdf.py} - the distance weight}

\item{ent_weight}{Argument for \code{update_hdf.py} - the entropy weight}

\item{dist_pow}{Argument for \code{update_hdf.py} - the distance power}

\item{obs_weight}{Argument for \code{update_hdf.py} - the observation weight}

\item{learning_rate}{Argument for \code{update_hdf.py}}

\item{training_steps}{Argument for \code{update_hdf.py}.}

\item{rng_seed}{Argument for \code{update_hdf.py}.}

\item{ebirdst_year}{Argument for \code{update_hdf.py}.}
}
\description{
Function to fit BirdFlow model by executing \code{update_hdf.py} in Python.
Arguments typically passed from \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}} via
\code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}}
}
\seealso{
\code{\link[=birdflow_modelfit_args_df]{birdflow_modelfit_args_df()}}, \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}
}
