% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{birdflow_modelfit_args_df}
\alias{birdflow_modelfit_args_df}
\title{Create a grid-expanded data.frame of model fit arguments. Output designed to be passed to \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}}
\usage{
birdflow_modelfit_args_df(params)
}
\arguments{
\item{params}{Parameters object, typically created during preprocessing}
}
\description{
Create a grid-expanded data.frame of model fit arguments. Output designed to be passed to \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}
}
\seealso{
\code{\link[=birdflow_modelfit]{birdflow_modelfit()}}, \code{\link[=batch_modelfit_wrapper]{batch_modelfit_wrapper()}}
}
