% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdFlowPipeline-package.R
\docType{package}
\name{BirdFlowPipeline-package}
\alias{BirdFlowPipeline}
\alias{BirdFlowPipeline-package}
\title{BirdFlowPipeline: Tools for fitting, selecting, and evaluating BirdFlow models on Unity}
\description{
Tools for fitting, selecting, and evaluating BirdFlow models on Unity.
}
\author{
\strong{Maintainer}: David Slager \email{slager@users.noreply.github.com} (\href{https://orcid.org/0000-0003-2525-2039}{ORCID})

Authors:
\itemize{
  \item Ethan Plunkett \email{plunkett@umass.edu} (\href{https://orcid.org/0000-0003-4405-2251}{ORCID})
}

Other contributors:
\itemize{
  \item BirdFlow-Science [copyright holder]
  \item National Science Foundation [funder]
}

}
\keyword{internal}
