% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_multisource_data.R
\name{combine_and_save_ground_truth_data}
\alias{combine_and_save_ground_truth_data}
\title{Combining the multsource data for one species}
\usage{
combine_and_save_ground_truth_data(species)
}
\arguments{
\item{species}{the species to process specified by species code}
}
\description{
Combining the multsource data for one species
}
