% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_data_combine_multisource.R
\name{combine_data_for_all_sp}
\alias{combine_data_for_all_sp}
\title{Combining the multsource data for all species}
\usage{
combine_data_for_all_sp(resolution = 100, max_n = 10000)
}
\description{
Combining the multsource data for all species
}
