% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a BirdFlow model}
\usage{
evaluate_model(path, track_info)
}
\arguments{
\item{path}{Path of BirdFlow model to evaluate}

\item{track_info}{Object produced from \code{\link[=make_tracks]{make_tracks()}}}
}
\value{
A list:
\itemize{
\item \code{df} 1-row data.frame of model descriptors and metrics
\item \code{obs} obs_df portion from original \code{track_info}
\item \code{int} = int_df portion from original \code{track_info}
}
}
\description{
Evaluate a BirdFlow model
}
\seealso{
\code{\link[=make_tracks]{make_tracks()}}, \code{\link[=batch_evaluate_models]{batch_evaluate_models()}}
}
