% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a BirdFlow model
This function could be user-customized}
\usage{
evaluate_model(bf_path, data, params)
}
\description{
Evaluate a BirdFlow model
This function could be user-customized
}
