% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a BirdFlow model}
\usage{
evaluate_model(
  bf,
  modelname,
  birdflow_intervals,
  birdflow_intervals_one_week,
  params
)
}
\arguments{
\item{bf}{A birdflow model object, often passed from \code{\link[=import_birdflow_and_evaluate]{import_birdflow_and_evaluate()}}}

\item{modelname}{A name for the model, often the basename of the model path}

\item{birdflow_intervals}{The BirdFlowIntervals object for evaluation}

\item{birdflow_intervals_one_week}{The one-week BirdFlowIntervals object for evaluation}

\item{params}{Standard params object}
}
\value{
A list:
\itemize{
\item \code{df} 1-row data.frame of model descriptors and metrics
}
}
\description{
Evaluate a BirdFlow model
}
\seealso{
\code{\link[=batch_evaluate_models]{batch_evaluate_models()}}, \code{\link[=import_birdflow_and_evaluate]{import_birdflow_and_evaluate()}}
}
