% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_efficiency.R
\name{get_job_efficiency}
\alias{get_job_efficiency}
\title{get the job efficiency of a completed job from slurm}
\usage{
get_job_efficiency(id)
}
\arguments{
\item{id}{The slurm job ID including an array suffix e.g. "_1" if
appropriate.}
}
\value{
A list with items:
\item{\code{job_id}}{slurm job id - without array suffix}
\item{\code{array_job_id}}{slurm job id with array suffix}
\item{\code{state}}{The job state}
\item{\code{cores}}{How many cores did the job use.}
\item{\code{cpu_utilized}}{How much CPU time was used.}
\item{\code{cpu_efficiency}}{CPU efficiency}
\item{\code{walltime}}{ Job wall-clock time (ellapsed time) in h:m:s format}
\item{\code{memory_utilized}}{Memory utilized in human readible format with (varying) units}
\item{\code{memory_efficiency}}{Memory efficiency.}
\item{\code{mem_gb}}{Utilized memory in GiB (bytes * 1024^3)}
\item{\code{wall_min}}{Job wall-clock time in decimal minutes}
}
\description{
\code{get_job_efficiency()} calls the shell command \code{"seff"} on the job ID
via \code{\link[batchtools:runOSCommand]{batchtools::runOSCommand()}}, which uses ssh to connect to a login
node before running the command.  The resulting output is then parsed
into an R list.
}
\seealso{
\code{\link[batchtools:findJobs]{batchtools::findJobs()}}
}
