% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_loader.R
\name{get_transitions}
\alias{get_transitions}
\title{Functions to get ground truth routes and transitions for model tuning and validation, combining tracking, Motus, and banding.}
\usage{
get_transitions(loader)
}
\arguments{
\item{loader}{TransitionsLoader class}
}
\value{
list of combined_routes_data, interval_obj, interval_one_week_obj
}
\description{
Functions to get ground truth routes and transitions for model tuning and validation, combining tracking, Motus, and banding.
}
