% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flux_report.R
\name{make_flux_report}
\alias{make_flux_report}
\title{Make flux report}
\usage{
make_flux_report(hdf, flux, anim, html, prevlink, nextlink)
}
\arguments{
\item{hdf}{path to HDF5 model file}

\item{flux}{path to flux RDS file}

\item{anim}{path to abundance and flux animation GIF}

\item{html}{path to output html file}

\item{prevlink}{relative path to the previous species report or NULL if
there is none.}

\item{nextlink}{relative path to the next species report or NULL if none.}
}
\value{
Nothing is returned, a report is created at \code{html}
}
\description{
Note: I ran this on unity after using \code{devtools::load_all()} it
seemed to be writing temporary files to the inst/rmd directory.  This
suggests that it might not work with an installed version of BirdFlowPipeline.
If that's the case we'd have to copy the Rmd from the package to a temorary
location before knitting.
}
\examples{
\dontrun{
base <- "/work/pi_drsheldon_umass_edu/birdflow_modeling/pipeline"
dir <- file.path(base, "/americas/acafly_100km")
html_dir <-file.path(base, "model_release_staging/flux/")
hdf <- file.path(
  dir, 
  "acafly_2022_100km_obs1.0_ent0.001924_dist0.008177_pow0.4167.hdf5")
flux <- file.path(dir, "acafly_flux.rds")
anim <- file.path(dir, "acafly_anim.gif")
html <- file.path(html_dir, "acafly.html")
prevlink <- NULL
nextlink <- "./adafly.html"
make_flux_report(hdf = hdf), 
                flux = flux, 
                anim = anim, 
                html = html,
                prevlink = prevlink, 
                nextlink = nextlink)
}
}
