% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rank_models}
\alias{rank_models}
\title{Rank the models in an eval_metrics data frame}
\usage{
rank_models(eval_metrics, params)
}
\arguments{
\item{eval_metrics}{The eval_metrics data frame often produced by
recombining results from \code{\link[=batch_evaluate_models]{batch_evaluate_models()}}}

\item{params}{The standard params list, which includes \code{model_selection} method}
}
\value{
Updated eval_metrics including desirability columns
}
\description{
Rank the models in an eval_metrics data frame
}
