% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy_crosswalk}
\alias{taxonomy_crosswalk}
\title{data.frame for matching USGS banding taxa to ebirdst taxa}
\format{
\subsection{\code{taxonomy_crosswalk}}{

A data.frame with 1072 obs. of 8 variables, made from eBird 2020 Taxonomy, USGS banding dataset \code{species.csv} table, and manual file \code{taxonomic_join_overrides.csv}.
\describe{
\item{BBL_SPECIES_ID}{4-digit USGS BBL species code}
\item{BBL_SPECIES_NAME}{USGS BBL common name}
\item{BBL_SCI_NAME}{USGS BBL scientific name}
\item{EBIRD_SPECIES_CODE}{6-letter primary key for eBird, sometimes with trailing number}
\item{EBIRD_PRIMARY_COM_NAME}{eBird common name}
\item{EBIRD_CATEGORY}{species, hybrid, slash, spuh, or NA}
\item{EBIRDST_CODE}{same as eBird species code, consider removing one column}

}
}
}
\source{
\url{https://dx.doi.org/10.5066/P9BSM38F}

\url{https://www.birds.cornell.edu/clementschecklist/download/}
}
\usage{
taxonomy_crosswalk
}
\description{
data.frame for matching USGS banding taxa to ebirdst taxa
}
\keyword{datasets}
